White space defect when parsing adjacent entities
Add unit tests to check for white space between XML entities.  One
test is the original example of two adjacent entities. An additional
test case *has* white space between entities, and is just an
affirmation of the existing behavior, and should be preserved.

This bug pops up when the preserveWS option is either enabled or
disabled for ConstructingParser.  This doubles the number of unit
tests to four.

The fifth unit test, listed first, of adjacent text characters is
rhetoric, and doesn't need to be included in the final merge.  The
previous sentence shouldn't be merged either.

[error] Test XMLTest.preserveNoSpaceBetweenEntitiesOptionEnabled failed:
[error]   expected: <<div>&lt;[]&lt;</div>>
[error]   but was: <<div>&lt;[ ]&lt;</div>>
[error] Test XMLTest.preserveNoSpaceBetweenEntitiesOptionDisabled failed:
[error]   expected: <<div>&lt;[]&lt;</div>>
[error]   but was: <<div>&lt;[ ]&lt;</div>>